Cloud DVD restoration: arbitrary-disc/NTSC support + money-safety hardening
Two bodies of work to make the two remaining DVD weddings (Gulfraz NTSC,
Mo PAL-TFF) restorable on local + rented RunPod GPUs. The second was driven
by an independent xhigh code review run BEFORE the first paid multi-pod run;
its blockers are fixed and unit-verified without a GPU/RunPod.

Arbitrary-disc + per-disc correctness (Phase 1)
- settings: DISC_REGISTRY allow-list carries slug/collection/standard/field_order;
  catalog ingests any staged ISO, not the two hardcoded ones (migration 007).
- pipeline_v3.sh: stage-1 deinterlace parameterised via DEINT_* (defaults = today's
  PAL/BFF); services.deinterlace_profile derives NTSC (59.94/640x480/smpte170m) and
  TFF per disc; frame-count math generalised to real output_fps.

Cross-GPU correctness (review blocker)
- slicer takes the job's real output_fps instead of hardcoding 50, so NTSC (59.94)
  cloud slices cut on the correct frame. The old default shipped shifted footage
  that frame-count validation could not catch (only unit 0 was correct — why the
  single-unit smoke test passed). test_slice_equivalence.sh now derives fps from
  the input and proves interior NTSC units are frame-exact.

Money-safety (review blockers)
- fleet: a _watchdog thread enforces WEDDING_CLOUD_SPEND_CAP_USD WHILE pods run
  (not just at bring-up) and tears the fleet down at the cap; pod_ttl_s enforced,
  refreshed per completed unit via hand_back().
- fleet._terminate marks a pod 'terminated' only when confirmed; an unconfirmed
  delete stays 'terminating' (spend keeps counting, reaper finishes). terminate_pod
  treats only 404/not-found as success (a 400 no longer masquerades as gone).
  _reconcile_cloud_pods reconciles the ledger against RunPod's live truth.
- idempotent create: the POST is no longer transport-retried; _create_pod adopts an
  existing same-name pod, so a lost response can't rent a billing twin (<=3/slot).
- bounded auto-resume: jobs.auto_resume_count caps recovery at 3 then parks 'failed';
  reset on human start/resume/retry or completion (migration 008).
- reaper: balance() decoupled from the delete path; --loop + --max-age-hours; ships
  as an age-filtered systemd timer (wedding-reaper.timer, >4h) that catches true
  orphans without touching a healthy in-progress fleet.
- executor retries a unit download 3x before discarding already-restored (paid) work;
  cloud pause lets in-flight paid units finish (only cancel/shutdown hard-abort);
  cloud_views counts creating/terminating pods as billing.

Concurrency
- per-job lease (jobs.lease_expires_at) renewed by a background thread; recovery
  requeues only stale-lease jobs, so a second (cloud) worker can no longer requeue
  and double-run the job a live (local) worker still owns.

Verified with scratch-DB + fake-client + synthetic-ffmpeg proofs; not yet
re-verified end-to-end against real pods.

+++ webapp/worker/slicer.py
@@ -24,7 +24,7 @@ def slice_unit(
     cap: int,
     dest: Path,
     *,
-    fps: int = 50,
+    fps: float = 50.0,
     ffmpeg_image: str = "linuxserver/ffmpeg:latest",
     data_dir: Path | None = None,
     use_docker: bool = True,
@@ -51,6 +51,15 @@ def slice_unit(
     dest.parent.mkdir(parents=True, exist_ok=True)
     # keep the real extension so ffmpeg can pick a muxer: name.partial.mkv
     part = dest.parent / f"{dest.stem}.partial{dest.suffix}"
+    # `-ss` seeks by TIME, and stage-1 is intra-only FFV1, so the seek is
+    # frame-exact ONLY when `fps` matches the stage-1 output frame rate. The
+    # caller MUST pass the job's real output fps (50 for PAL, 60000/1001≈59.94
+    # for NTSC): frame `skip` sits at PTS skip/fps, and using the wrong fps
+    # lands on the wrong frame. For NTSC at the default 50 that error is ~19% of
+    # `skip` and grows every unit, yet frame-count validation still passes
+    # (interior slices still hold `cap` frames) — so the whole cloud run would
+    # silently restore shifted footage. The 6-decimal timestamp is well within
+    # half a frame of the true PTS, so it snaps to exactly frame `skip`.
     ss = f"{skip / fps:.6f}"
 
     if use_docker:
